support separate add and remove rows
