+++ frontend/src/stories/features/crawl-workflows/CustomBehaviorsTable.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/web-components";
 import { html } from "lit";
 
-import { argTypes } from "../excludeBtrixElementProperties";
+import { argTypes } from "../excludeContainerProperties";
 
 import type { CustomBehaviorsTable } from "@/features/crawl-workflows/custom-behaviors-table";
