+++ frontend/src/stories/features/excludeContainerProperties.ts
@@ -1,6 +1,8 @@
 /**
- * Exclude `BtrixElement` properties from story controls
+ * Exclude `BtrixElement` and `StorybookOrg` properties from story controls
  */
+import { StorybookOrg } from "../decorators/orgDecorator";
+
 import { BtrixElement } from "@/classes/BtrixElement";
 
 const controlOpts = { table: { disable: true } };
@@ -12,6 +14,12 @@ const argTypes: Record<string, typeof controlOpts> = {
 };
 
 Object.getOwnPropertyNames(BtrixElement.prototype).forEach((prop) => {
+  if (prop === "constructor") return;
+
+  argTypes[prop] = controlOpts;
+});
+
+Object.getOwnPropertyNames(StorybookOrg.prototype).forEach((prop) => {
   if (prop === "constructor") return;
 
   argTypes[prop] = controlOpts;
